fix: Anotações '@Columns' retiradas, tipo ID alterado de Integer para Long

+++ src/main/java/com/restaurant/repositories/ProductsRepository.java
@@ -4,7 +4,7 @@ import com.restaurant.models.ProductsEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-
 @Repository
-public interface ProductsRepository extends JpaRepository<ProductsEntity, Integer> {
+public interface ProductsRepository
+        extends JpaRepository<ProductsEntity, Long> {
 }
